pyproject.toml
[project]
dependencies = [
    "django",
    "django-auditlog==2.2.0",
    "django-braces",
    "django-debug-toolbar",
    "django-environ",
    "django-extensions",
    "django-filter",
    "django-handyhelpers",
    "django-hostutils",
    "django-userextensions",
    "django-pygwalker",
    "django-signalcontrol",
    "djangorestframework",
    "djangorestframework-filters==1.0.0.dev0",
    "drf-flex-fields",
    "drf-spectacular",
    "drf-renderer-xlsx",
]
description = "django project created from djangoaddicts django_project_template"
dynamic = ["version"]
keywords = ["django"]
license = {file = "LICENSE"}
name = "my_django_project"
readme = "README.md"
requires-python = ">=3.8"


[project.optional-dependencies]
dev = [
    "bandit",
    "black",
    "coveralls",
    "isort",
    "model-bakery",
    "mypy",
    "mypy-extensions",
    "ruff",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "radon",
    "safety",
    "types-python-dateutil",
    "typing_extensions",
]

[tool.bandit]
exclude_dirs = ["venv", "django_project/tests"]


[tool.black]
line-length = 120


[tool.coverage.report]
show_missing = true


[tool.coverage.run]
branch = true
source = ["django_project"]
omit = [
    "django_project/manage.py", 
    "django_project/core/asgi.py",
    "django_project/core/wsgi.py", 
    "django_project/*/scripts/*",
    "django_project/tests/*", 
    ]


[tool.flake8]
ignore = "E203,E266,H106,H904,E133"
max-line-length = 120
max-complexity = 25
hang-closing = true
exclude = "django_project/manage.py, django_project/*/scripts"


[tool.pylint.format]
max-line-length = "120"

[tool.pylint.MASTER]
ignore-paths = ["django_project/manage.py", "django_project/tests", "django_project/store/migrations", "django_project/*/scripts", "django_project/*/local_test"]
exit-zero = true

[tool.pylint.'MESSAGES CONTROL']
disable = "R0903,R0913,R0901,W0613,W0718,C0114,C0115,C0103,R1725"


[tool.pytest.ini_options]
addopts = "-s -v -x --strict-markers -m 'not extra' --cov=django_project"
testpaths = ["django_project"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]


[tool.ruff]
line-length = 120
exclude = ["django_project/manage.py", "django_project/tests", "django_project/*/migrations", "django_project/*/scripts", "django_project/*/local_test"]
